separate view directory
