+++ lib/page/details/detail_page.dart
