+++ lib/page/details/detail_page_argument.dart
