+++ lib/page/root/top_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
-import 'package:flutter_study_day7/page/year_tab_page.dart';
+import 'package:flutter_study_day7/page/root/year_tab_page.dart';
 
 class TabInfo {
   int year;
